Remove Snapp code
