+++ README.md
@@ -40,30 +40,21 @@ If you are building for a local development network, ganache has to be running l
 yarn run ganache # start a development network (blocking)
 ```
 
-In case only the StableX or only the SnappAuction should be deployed, you can add the flags `--onlyMigrateStableX` or `--onlyMigrateSnappAuction` to the deployment command.
-
-E.g:
-
-```
- npx truffle migrate --onlyMigrateStableX=true
-```
-
 3. Verify the contracts for some cool Etherscan.io goodies (see below for more help)
 
 ```sh
-npx truffle run verify SnappAuction --network $NETWORKNAME
 npx truffle run verify BatchExchange --network $NETWORKNAME
 ```
 
 4. List some default tokens on the StableX exchange
 
 ```sh
-npx truffle exec scripts/stablex/add_token_list.js --network $NETWORKNAME
+npx truffle exec scripts/add_token_list.js --network $NETWORKNAME
 ```
 
 ## Verifying Contracts
 
-In order to verify a contract on Etherscan.io, you need to first creat an account and an API key
+In order to verify a contract on Etherscan.io, you need to first create an account and an API key
 
 1. Navigate to https://etherscan.io/myapikey
 2. Login or create an account
@@ -98,29 +89,29 @@ The following script deploys a simple market maker order and a necessary owl ord
 
 ```sh
 # Get token ID of DAI
-npx truffle exec scripts/stablex/invokeViewFunction.js 'tokenAddressToIdMap' '0x5592EC0cfb4dbc12D3aB100b257153436a1f0FEa' --network rinkeby
+npx truffle exec scripts/invokeViewFunction.js 'tokenAddressToIdMap' '0x5592EC0cfb4dbc12D3aB100b257153436a1f0FEa' --network rinkeby
 
 # Export the resulting token ID
 export TOKEN_ID_DAI=[Result from last call]
 
 # Get token ID of TrueUSD
-npx truffle exec scripts/stablex/invokeViewFunction.js 'tokenAddressToIdMap' '0x0000000000085d4780B73119b644AE5ecd22b376' --network rinkeby
+npx truffle exec scripts/invokeViewFunction.js 'tokenAddressToIdMap' '0x0000000000085d4780B73119b644AE5ecd22b376' --network rinkeby
 
 # Export the resulting token ID
 export TOKEN_ID_TUSD=[Result from last call]
 
 # Make sure that the users have deposited sufficient funds into the exchange
 # Please be aware that the specified amounts are multiples of 10**18
-npx truffle exec scripts/stablex/deposit.js --accountId=0 --tokenId=0 --amount=30 --network rinkeby&& \
-npx truffle exec scripts/stablex/deposit.js --accountId=0 --tokenId=$TOKEN_ID_TUSD --amount=100 --network rinkeby
+npx truffle exec scripts/deposit.js --accountId=0 --tokenId=0 --amount=30 --network rinkeby&& \
+npx truffle exec scripts/deposit.js --accountId=0 --tokenId=$TOKEN_ID_TUSD --amount=100 --network rinkeby
 
 # Place  market-maker order in current auction
 # This simulates a strategy expected from market makers: trading stable coins against each other
 # with a spread of 0.02 percent
-npx truffle exec scripts/stablex/place_order.js --accountId=0 --buyToken=$TOKEN_ID_DAI --sellToken=$TOKEN_ID_TUSD --minBuy=1000 --maxSell=998 --validFor=20 --network rinkeby
+npx truffle exec scripts/place_order.js --accountId=0 --buyToken=$TOKEN_ID_DAI --sellToken=$TOKEN_ID_TUSD --minBuy=1000 --maxSell=998 --validFor=20 --network rinkeby
 
 # Place owl token order for the fee mechanism
-npx truffle exec scripts/stablex/place_order.js --accountId=0 --buyToken=$TOKEN_ID_DAI --sellToken=0 --minBuy=1000 --maxSell=1000 --validFor=20 --network rinkeby
+npx truffle exec scripts/place_order.js --accountId=0 --buyToken=$TOKEN_ID_DAI --sellToken=0 --minBuy=1000 --maxSell=1000 --validFor=20 --network rinkeby
 
 ```
 
@@ -128,17 +119,17 @@ Then, the market order can be place, after switching to another account. Usually
 
 ```sh
 # Deposit funds into exchange:
-npx truffle exec scripts/stablex/deposit.js --accountId=0 --tokenId=$TOKEN_ID_DAI --amount=100 --network rinkeby
+npx truffle exec scripts/deposit.js --accountId=0 --tokenId=$TOKEN_ID_DAI --amount=100 --network rinkeby
 
 # Place market order with 1/2 limit-price
-npx truffle exec scripts/stablex/place_order.js --accountId=1 --buyToken=$TOKEN_ID_TUSD --sellToken=$TOKEN_ID_DAI --minBuy=500 --maxSell=1000 --validFor=5 --network rinkeby
+npx truffle exec scripts/place_order.js --accountId=1 --buyToken=$TOKEN_ID_TUSD --sellToken=$TOKEN_ID_DAI --minBuy=500 --maxSell=1000 --validFor=5 --network rinkeby
 ```
 
 Now, the market can be inspected by:
 
 ```sh
 # view the market status:
-npx truffle exec scripts/stablex/get_auction_elements.js --network rinkeby
+npx truffle exec scripts/get_auction_elements.js --network rinkeby
 
 ```
