+++ package.json
@@ -13,7 +13,6 @@
     "build": "truffle build && yarn run networks-reset && yarn run ts",
     "test": "truffle test && yarn run test-ts",
     "test-stablex": "yarn run test-js test/stablex/*",
-    "test-snapp": "yarn run test-js test/snapp/*",
     "test-js": "truffle test",
     "test-ts": "mocha -r ts-node/register test/models/*.spec.ts",
     "coverage": "truffle run coverage --network coverage --file \"test/**/*js\"",
@@ -27,12 +26,12 @@
     "prettier:solidity": "prettier --write 'contracts/**/*.sol'",
     "prettier:js": "prettier --write './**/*.js'",
     "pretty-check": "prettier --check 'contracts/**/*.sol' && prettier --check './**/*.js'",
-    "spread-orders": "npx truffle exec scripts/stablex/place_spread_orders.js",
+    "spread-orders": "npx truffle exec scripts/place_spread_orders.js",
     "ts": "yarn run ts-contracts && yarn run ts-check && yarn run tsc:commonjs && yarn run tsc:esm",
     "tsc:commonjs": "rimraf typescript/common && tsc --outDir typescript/common",
     "tsc:esm": "rimraf typescript/esm && tsc -m es6 -t esnext --outDir typescript/esm",
     "ts-check": "tsc --esModuleInterop --noEmit src/index.d.ts",
-    "ts-contracts": "typechain --target web3-v1 --outDir build/types \"build/contracts/{BatchExchange,SnappAuction,BatchExchangeViewer}.json\"",
+    "ts-contracts": "typechain --target web3-v1 --outDir build/types \"build/contracts/{BatchExchange,BatchExchangeViewer}.json\"",
     "ganache": "ganache-cli -h 0.0.0.0 --gasLimit 8e6 --deterministic"
   },
   "repository": {
@@ -45,9 +44,6 @@
     "build/contracts/EpochTokenLocker.json",
     "build/contracts/IERC20.json",
     "build/contracts/Migrations.json",
-    "build/contracts/SnappAuction.json",
-    "build/contracts/SnappBase.json",
-    "build/contracts/SnappBaseCore.json",
     "build/contracts/TokenConservation.json",
     "build/types",
     "typescript/common",
