+++ scripts/add_token.js
