+++ scripts/add_token_list.js
