+++ scripts/close_auction.js
