+++ scripts/ensure_owl_liquidity.js
@@ -2,7 +2,7 @@ const BatchExchange = artifacts.require("BatchExchange")
 
 const BN = require("bn.js")
 const { maxUint32, sendLiquidityOrders } = require("./utilities")
-const { getOrdersPaginated } = require("../../src/onchain_reading")
+const { getOrdersPaginated } = require("../src/onchain_reading")
 
 const MINIMAL_LIQUIDITY_FOR_OWL = new BN(10).pow(new BN(17))
 const SELL_ORDER_AMOUNT_OWL = new BN(10).pow(new BN(18)).mul(new BN(5))
