+++ scripts/invokeViewFunction.js
@@ -1,5 +1,5 @@
 const BatchExchange = artifacts.require("BatchExchange")
-const { invokeViewFunction } = require("../script_utilities.js")
+const { invokeViewFunction } = require("./utilities.js")
 
 module.exports = async (callback) => {
   await invokeViewFunction(BatchExchange, callback)
