+++ scripts/setup_environment.js
