+++ scripts/transitive_orderbook.js
@@ -1,9 +1,9 @@
 const BatchExchangeViewer = artifacts.require("BatchExchangeViewer")
-const { getOpenOrdersPaginated } = require("../../src/onchain_reading.js")
+const { getOpenOrdersPaginated } = require("../src/onchain_reading.js")
 const BN = require("bn.js")
 
-const { Orderbook, Offer, transitiveOrderbook } = require("../../typescript/common/orderbook.js")
-const { Fraction } = require("../../typescript/common/fraction.js")
+const { Orderbook, Offer, transitiveOrderbook } = require("../typescript/common/orderbook.js")
+const { Fraction } = require("../typescript/common/fraction.js")
 
 const argv = require("yargs")
   .option("sellToken", {
