+++ src/index.js
@@ -9,7 +9,6 @@
 module.exports = {
   BatchExchange: require("../build/contracts/BatchExchange.json"),
   BatchExchangeViewer: require("../build/contracts/BatchExchangeViewer.json"),
-  SnappAuction: require("../build/contracts/SnappAuction.json"),
   ...require("../typescript/common/fraction.js"),
   ...require("../typescript/common/orderbook.js"),
   ...require("./encoding.js"),
