+++ test/stablex/epoch_token_locker.js
@@ -4,7 +4,7 @@ const MockContract = artifacts.require("MockContract")
 const ERC20Interface = artifacts.require("ERC20")
 
 const truffleAssert = require("truffle-assertions")
-const { closeAuction } = require("../../scripts/stablex/utilities")
+const { closeAuction } = require("../../scripts/utilities")
 
 contract("EpochTokenLocker", async (accounts) => {
   const [user_1, user_2] = accounts
