+++ test/stablex/regression_tests.js
@@ -6,7 +6,7 @@ const IterableAppendOnlySet = artifacts.require("IterableAppendOnlySet")
 const BN = require("bn.js")
 const truffleAssert = require("truffle-assertions")
 
-const { closeAuction } = require("../../scripts/stablex/utilities.js")
+const { closeAuction } = require("../../scripts/utilities.js")
 
 const { solutionSubmissionParams, basicTrade, utilityOverflow } = require("../resources/examples")
 const { makeDeposits, placeOrders, setupGenericStableX } = require("./stablex_utils")
