+++ test/stablex/stablex_large_example.js
@@ -1,4 +1,4 @@
-const { closeAuction } = require("../../scripts/stablex/utilities.js")
+const { closeAuction } = require("../../scripts/utilities.js")
 const { solutionSubmissionParams, largeRing30 } = require("../resources/examples")
 const { makeDeposits, placeOrders, setupGenericStableX } = require("./stablex_utils")
